Session 7 task/requirements.txt
flask==3.0.2
google-generativeai==0.3.2
youtube-search==2.1.2
youtube-transcript-api==0.6.2
numpy==1.24.3
faiss-cpu==1.7.4
requests==2.31.0
pydantic==2.6.3
markitdown==0.1.3
streamlit==1.31.0
mcp==0.4.0
python-dotenv==1.0.1
transformers==4.38.2 